Move folders to live outside components folder
