+++ web-registry/src/pages/Admin.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
 
-import { NavBar } from '../organisms';
-import isAdmin from '../../lib/admin';
-import getRegistryUrl from '../../lib/registryUrl';
+import { NavBar } from '../components/organisms';
+import isAdmin from '../lib/admin';
+import getRegistryUrl from '../lib/registryUrl';
 
 function Admin(): JSX.Element {
   const { user } = useAuth0();
