+++ web-registry/src/pages/BuyerCreate.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { makeStyles, Theme } from '@material-ui/core/styles';
 import { useMutation } from '@apollo/client';
-// import { gql } from '@apollo/client';
+
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import FormLabel from '@material-ui/core/FormLabel';
@@ -14,8 +14,8 @@ import { loader } from 'graphql.macro';
 import Title from 'web-components/lib/components/title';
 import Geocoder from 'web-components/lib/components/map/Geocoder';
 
-const CREATE_USER = loader('../../graphql/ReallyCreateUser.graphql');
-const CREATE_USER_ORGANIZATION = loader('../../graphql/CreateUserOrganization.graphql');
+const CREATE_USER = loader('../graphql/ReallyCreateUser.graphql');
+const CREATE_USER_ORGANIZATION = loader('../graphql/CreateUserOrganization.graphql');
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
