+++ web-registry/src/pages/Credits.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { PurchasedCredits, purchasedCredits } from '../../mocks';
-import { UserCredits } from '../organisms';
+import { PurchasedCredits, purchasedCredits } from '../mocks';
+import { UserCredits } from '../components/organisms';
 
 function Credits(): JSX.Element {
   let { userId } = useParams<{ userId: string }>();
