+++ web-registry/src/pages/GettingStarted.tsx
