+++ web-registry/src/pages/Projects.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { projects } from '../../mocks';
-import { ProjectList } from '../organisms';
+import { projects } from '../mocks';
+import { ProjectList } from '../components/organisms';
 
 function Projects(): JSX.Element {
   return <ProjectList projects={projects} />;
